added CommentRepository, CommentNotFoundException Added methods in CommentServiceImpl

+++ src/main/java/com/skyteam/shopping_area/repository/AdsRepository.java
@@ -3,5 +3,5 @@ package com.skyteam.shopping_area.repository;
 import com.skyteam.shopping_area.model.Ad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdsRepository extends JpaRepository<Ad, Long> {
+public interface AdsRepository extends JpaRepository<Ad, Integer> {
 }
